+++ packages/common/src/index.ts
@@ -32,6 +32,7 @@ export {
   valueOrNA,
   zodBuilder,
   zodErrorToReadable,
+  checkIsNonEmptyArrayOfNonEmptyStrings,
   getCountries,
   getCountryStates,
   getFullCountryNameByCode,
@@ -40,6 +41,11 @@ export {
 } from './utils';
 
 export * from './utils/collection-flow';
+export type {
+  TCollectionFlow,
+  TCollectionFlowState,
+  TCollectionFlowStep,
+} from './utils/collection-flow';
 
 export type { IErrorWithMessage } from './utils';