+++ services/workflows-service/src/workflow/schemas/zod-schemas.ts
@@ -12,13 +12,13 @@ export const ConfigSchema = z
     isLegacyReject: z.boolean().optional(),
     isLockedDocumentCategoryAndType: z.boolean().optional(),
     isManualCreation: z.boolean().optional(),
-    isDemo: z.boolean().optional(),
     isExample: z.boolean().optional(), // OSS only
     language: z.string().optional(),
     supportedLanguages: z.array(z.string()).optional(),
     subscriptions: z.array(SubscriptionSchema).optional(),
     completedWhenTasksResolved: z.boolean().optional(),
     workflowLevelResolution: z.boolean().optional(),
+    isCollectionFlowPageRevisionEnabled: z.boolean().optional(),
     allowMultipleActiveWorkflows: z.boolean().optional(),
     initialEvent: z.string().optional(),
     availableDocuments: z.array(z.object({ category: z.string(), type: z.string() })).optional(),
@@ -64,6 +64,7 @@ export const ConfigSchema = z
     hasUboOngoingMonitoring: z.boolean().optional(),
     maxBusinessReports: z.number().nonnegative().optional(),
     isMerchantMonitoringEnabled: z.boolean().optional(),
+    isOngoingMonitoringEnabled: z.boolean().optional(),
     isDemoAccount: z.boolean().optional(),
     withQualityControl: z.boolean().optional(),
     disableBusinessSyncToUnifiedApi: z.boolean().optional(),
@@ -106,6 +107,8 @@ export const CustomerConfigSchema = z.object({
   hideCreateMerchantMonitoringButton: z.boolean().optional(),
   isExample: z.boolean().optional(),
   isMerchantMonitoringEnabled: z.boolean().optional(),
+  isOngoingMonitoringEnabled: z.boolean().optional(),
+  isDemo: z.boolean().optional(),
   maxBusinessReports: z.number().optional(),
 });