+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @ballerine/backoffice-v2
 
+## 0.7.126
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.86
+  - @ballerine/ui@0.7.126
+  - @ballerine/workflow-browser-sdk@0.6.108
+  - @ballerine/workflow-node-sdk@0.6.108
+  - @ballerine/react-pdf-toolkit@1.2.99
+
+## 0.7.125
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.85
+  - @ballerine/ui@0.7.125
+  - @ballerine/workflow-browser-sdk@0.6.107
+  - @ballerine/react-pdf-toolkit@1.2.98
+  - @ballerine/workflow-node-sdk@0.6.107
+
+## 0.7.124
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.124
+  - @ballerine/react-pdf-toolkit@1.2.97
+
 ## 0.7.123
 
 ### Patch Changes