+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,33 @@
 # kyb-app
 
+## 0.3.155
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.86
+  - @ballerine/ui@0.7.126
+  - @ballerine/workflow-browser-sdk@0.6.108
+
+## 0.3.154
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.85
+  - @ballerine/ui@0.7.125
+  - @ballerine/workflow-browser-sdk@0.6.107
+
+## 0.3.153
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/ui@0.7.124
+
 ## 0.3.152
 
 ### Patch Changes