+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/headless-example
 
+## 0.3.107
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.86
+  - @ballerine/workflow-browser-sdk@0.6.108
+
+## 0.3.106
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.85
+  - @ballerine/workflow-browser-sdk@0.6.107
+
 ## 0.3.105
 
 ### Patch Changes