+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.2.99
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.126
+
+## 1.2.98
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.125
+
+## 1.2.97
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.124
+
 ## 1.2.96
 
 ### Patch Changes