+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ballerine/ui
 
+## 0.7.126
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.86
+
+## 0.7.125
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.85
+
+## 0.7.124
+
+### Patch Changes
+
+- Bump
+
 ## 0.7.123
 
 ### Patch Changes