+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ballerine/workflow-core
 
+## 0.6.108
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.86
+
+## 0.6.107
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.85
+
 ## 0.6.106
 
 ### Patch Changes