+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ballerine/workflows-service
 
+## 0.7.115
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.86
+  - @ballerine/workflow-core@0.6.108
+  - @ballerine/workflow-node-sdk@0.6.108
+
+## 0.7.114
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.85
+  - @ballerine/workflow-core@0.6.107
+  - @ballerine/workflow-node-sdk@0.6.107
+
 ## 0.7.113
 
 ### Patch Changes